feat: add close-when-stale action (#2864)
This PR introduces a new GitHub Actions swmansion-bot feature - `close-when-stale`. The code of this action can be found here: https://github.com/software-mansion-labs/swmansion-bot/tree/main/close-when-stale.

This action runs daily and closes issues that have a specified label (here `close-when-stale`) with no activity for some time (here `20` days).

The label is automatically removed when activity on an issue from someone outside of the organization is detected.